requirements.txt
streamlit
openai
fastapi==0.119.1
uvicorn==0.38.0
python-dotenv==1.1.1
langchain==1.0.1
langchain-openai==1.0.0
langchain-community==0.4
langchain-chroma==1.0.0
langdetect==1.0.9
openai==1.109.1
chromadb==1.1.1
requests==2.32.5
pypdf==6.2.0